# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Punjab Blood Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new public user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAuthResponse"
        "409":
          description: Email already exists
  /auth/user/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login for public users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAuthResponse"
        "401":
          description: Invalid credentials

  # --- USER PROFILE ---
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get own user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update own user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user/change-password:
    post:
      operationId: changeUserPassword
      tags: [user]
      summary: Change user password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"

  # --- USER: BLOOD REQUESTS ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
...
    AuthMessage:
      type: object
      required: [message]
      properties:
        message: { type: string }
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }

    # --- HOSPITAL SELF-REGISTRATION ---
...
    UserRegistration:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        phone: { type: string }
        bloodGroup: { type: string }
        city: { type: string }
        district: { type: string }
    UserAuthResponse:
      type: object
      required: [token, role, user]
      properties:
        token: { type: string }
        role: { type: string }
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, name, email, isBlocked, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        bloodGroup: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        district: { type: ["string", "null"] }
        isBlocked: { type: boolean }
        createdAt: { type: string }
    UserProfileUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        bloodGroup: { type: string }
        city: { type: string }
        district: { type: string }

    # --- BLOOD REQUESTS ---
...